drag and move with translate now c'est tout le s ajout du resize de shape

+++ src/main/java/xshape/eventObserver/Observer.java
@@ -1,10 +1,12 @@
 package xshape.eventObserver;
 
+import java.awt.geom.Point2D;
+
 import xshape.shapes.Shape;
 
 public interface Observer {
 
     void updateSelectedShape(Shape shape);
 
-    void updateShapePosition(Shape shape, double x, double y);
+    void updateShapePosition(Shape shape, int x, int y);
 }
